+++ hail/src/main/scala/is/hail/shuffler/Wire.scala
@@ -4,20 +4,20 @@ import is.hail.HailLSM
 import is.hail.annotations.Region
 import is.hail.expr.ir.IRParser
 import is.hail.expr.types.encoded.EType
-import is.hail.expr.types.physical.{ PStruct, PType }
-import is.hail.expr.types.virtual.{ TStruct, Type }
+import is.hail.expr.types.physical.{PStruct, PType}
+import is.hail.expr.types.virtual.{TStruct, Type}
 import is.hail.io.TypedCodecSpec
 import is.hail.rvd.AbstractRVDSpec
 import java.io._
 import java.net._
 import java.security.KeyStore;
 import java.util.UUID
-import java.util.concurrent.{ ConcurrentSkipListMap, Executors }
+import java.util.concurrent.{ConcurrentSkipListMap, Executors}
 import javax.net._
 import javax.net.ssl._
 import javax.security.cert.X509Certificate;
-import org.json4s.jackson.{ JsonMethods, Serialization }
-import com.fasterxml.jackson.core.{ JsonGenerator, JsonParser }
+import org.json4s.jackson.{JsonMethods, Serialization}
+import com.fasterxml.jackson.core.{JsonGenerator, JsonParser}
 
 import scala.annotation.switch
